+++ Models/ViewModel/SalesOrderItemViewModel.cs
@@ -7,9 +7,9 @@ namespace IntegratedInventoryAndOrderManagementSystem.Models.ViewModel
 {
     public class SalesOrderItemViewModel
     {
-         public int Id { get; set; }
-    public int ProductId { get; set; }
-    public int Quantity { get; set; }
-    public double SellingPrice { get; set; } 
+        public int Id { get; set; }
+        public int ProductId { get; set; }
+        public int Quantity { get; set; }
+        public decimal SellingPrice { get; set; } 
     }
 }
